+++ CPlusPlus/Integrators.h
@@ -6,7 +6,7 @@
 
 class Integrators {
     public:
-    static std::vector<Body*> symplecticEuler(std::vector<Body*> bodies, double dt);
+    static std::vector<Body*> symplecticEuler(std::vector<Body*> bodies, float dt);
 
 };
